Move admin endpoints into separate files (#6308)
