+++ changelog.d/6308.misc
@@ -1,0 +1,1 @@
+Move admin endpoints into separate files. Contributed by Awesome Technologies Innovationslabor GmbH.
